Java GUI created with a code area, command area and a canvas area.
I have created the GUI for the application. The GUI has 3 areas the left hand side contains the code area for the user to write programs, the center contains the canvas area where the shapes will be drawn and the bottom contains the command area where commands can be entered.

The next step is to create the parser class which will validate the data entered by the user in the boxes and report any errors back to the calling class.