Code Validation Parser Completed
The parser functionality of validating the code has been completed. It works by first checking that the command is correct and then checks the paramaters based on the command entered. The single command area will run by default however the codeblock area will only run when run is typed in the command area and the submit butten is clicked.

Functionality to make the code block react like a traditional code editor has been added. This means the user can put tabs and spaces on a line and it will not impact the code running properly.

The next step is to create a JMenu and implement the ability for the user to save and load in code samples.