JMenu Functionality Added (Save and Load)
I have added the ability to save a file by getting the value from the code area into a string, creating a file and writing it with the string using a fileoutput writer. The file will save as a .txt.

I have also added the ability for the user to load a file which then will be read using the Files.readString() command. This in turn is sent to the codea area where the user then has access to the commands loaded.

The next step is to complete the drawing commands by executing specific methods after the parser has ensured they are valid.