Executable Commands, Factory Design Pattern Implemented and Design Improvements
Firstly I have implemented the factory design pattern, with the parent class being the Command and all subclassess being the commands the user can type in the code area. Each command has a syntax property (name) and the number of paramaters it expects. These properties are used when checking against the command the user has entered into the code area.

Each command class has a run method which executes that command by calling the appropriate method of the canvas pannel which is passed in as a paramater when the object is created.

For the GUI Layout changes I have added scrollbars to certain areas to allow for a more userfriendly experience. I have also created a new JPannel called Console which holds a JText area and which is passed to the parser which can then display error messages on it.

The next step is to implement the unit tests for this and next submission, test the commands for bugs and implement additional features which could include the ability to save the bitmap image aswell as the code and the ability to specifiy the width and height of the bitmap.