JUnit testing added and changes to Validation Class
Firstly I have made some changes to the validation parser class. Before this commit, the commas would be checked at the end to check for a syntax error. Now however the code makes use of them more effectively. The new method splits the users input by the fist occurence of a space. This leaves the command and the paramaters. We now try to split the paramaters again by the comma. These two arrays are then validated and recombined before the command is executed. This also allows for spaces between the commas, allowing the user more freedom with the editor.

Finally I have added some JUnit Tests for two Classes. The first Class. the CommandParser Class, checks the validation method for 4 diffrent cases, which are

- Command Correct
- Command Incorrect
- Paramaters Correct
- Paramaters Incorrect

This Class also contains tests for the second submission, checking that a variable can be set and retrived. And that a loop can be executed and that the loop has incremented a given variable.

The Command Factory Class has also been given testing. This test checks that the subclass returned is of the correct type depending on the string passed into the getclass method.

The next steps are to begin work on submission 2 looking at Variables and adding addtional features.