Working Var Keyword Added
The user is now able to add a variable to the program and call the variable in paramaters of other commands later on.

If the command is found not to be valid then we check to see if the user entered the var keyword. if they have then check then remove the var keyword and split the remaining values by the = sign. This leaves the name and value. A new variable object is then created passing in the key value pair.

This works by adding code after the command has been checked. If the command has not been found then it should check to see if the user has entered a variable. (this used to return false) The code then checks the variable array to see if theres any variables and if thats true then it will try to split the paramaters and execute to appropriate paramaters. The following functions have been added:

+ value
++
- value
--
* value
/ value

Functions work by calling the variable objects methods to get and retrive the values.