Loops Added
The addtion of loops has been made to the program. each loop has its own object which holds its start and end position and whether the loop is a completed loop with both a start and end tag.

A variable holds the number of loops to allow for multiple loops to be used within a program. once the syntax has been checked and the loop objects paramters updated the completed boolean is then checked to ensure that all loops are completed.

We then loop through the list of loops. for each single loop the commands between the start and end are added time to loop entered -1 this is because there is one instance already in the command. these commands are before the endloop tag of the stated loop.

Finally the loop tags are removed and the code is passed to other methods as if the user had written the looped commands themselves.

Variables also work with loops and are checked when the 3rd paramter of a loop is checked. Appropriate error messages for all errors have been provided in the gui console.