Changes to Loops
Loops now work correctly. when a loop is created its added to a loops array and when an end tag is found it is set to the last loop to be created. this loop is then added to the completed loops array and then removed from the original loops array.

The completed loops array is then sorted by hiarachy then by name (order created) this means that when looped backwards the most inner loop will be executed first and if there is two on the same level the one created first will be executed first. Loops now update the positions after any commands have been added and every loop is searched through to check if it meets the requirements to be updated.

Minor fixes:

Variables can now be set to another variable by checking if the paramter is equal to a variable then retriving its value.

Loops that create new variables will return an error if more than one instance would be created with the same name.

Loops that loop 0 times will now set the code between them to be deadcode which will be removed along with all tags at the end.