Single and Multi-Lined If Statements Added
The ability of the user to use single if statements have been added. When the parser first checks for the tags if it find the "if" keyword it then tries to split by the "then" keyword if the returned array has more than one value then its a single line as the then keyword was used if not then its a multiple lined variable. Single lined variables are then split accordingly and the start and end postion and comparison values of the statment is set in a new object along with the command associated with it. This is then added to a completed ifstatements arraylist.

Multiple lined if statements work the same way as loops with the first endif closing off the last added if in the ifstatements array. This loop then has its end tag set and its comparrison values, its then removed from its original array and placed into the completed ifstatements array.

Ifstatments are executed before loops with the singleifcommands property of the object determaining if the if is a single or multi lined if. The if is then checked appropriatley and if incorrect then for single lined ifs the entire line is set to deadcode to be removed at command execution. For multi-lined ifs if the statement is false then all lines from start to finish are set to deadcode. If the condition is true then the multilined variable start and finish tags will be set to deadcode to be removed and for a single lined if the line will be replaced with the contents of the singlelinedcommands property from the if object.

IF STATEMENTS DO NOT WORK PROPERLY IN LOOPS (although do work alone and work with variables)

The next step is to add working methods to the program both with and without paramters.