Improvements to the Factory Design Pattern and Singleton Design Pattern Added
The original use for the Factory Design Pattern was not correct as the values going into the factory were known. To show how it can be properly used when a command is ready to be executed the first paramater (The command name) is passed into the command factory and object is returned and then its run method is called.

The singleton design pattern has also been added to stop the program creating new objects each time a command has been entered. Each command class now has its constructor set to private and a static variable called instance which holds a null object of the class. The first time this method is called it will check to see if the object is null and create one. Any time after that it will return the object created previously. The factoryclass now calls the get instance instead of the constructor when trying to create an object.

The next stage is to implement the XML Documentation for the parser class and add methods to the program.