Additional Features Added
After attempting to add methods and not getting them to work correctly the implementation code has been removed. Two new features have been added in its place. The first of which is the ability to save the image created by the commands this is done by saving the bitmap to a jpg when the appropriate jmenuitem is clicked.

The second feature is comments, these allow the user to type a comment by including the // at the start of the line which is then ignored at execution time by being removed after being changed to deadcode.

The final step is to improve the commenting of the code for maintancence purpose including adding xml documentation