Final Added Features
I have added two more features one of which is a new command called randomcolor this will change the color to a random one and when used in loops makes some colorful artwork. I have also added a new jmenu and jmenu item which displays a list of commands when clicked on. This will make the user aware of the syntax of the language. Minor bug fixes have been added allowing spaces between diffren't parameters for some commands that didn't before.